+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `rector/rector` with our default config and `mds-rector` bin script
+- Update Packages
+  - `phpmd/phpmd`
+  - `phpstan/phpstan`
+  - `symplify/phpstan-rules`
+  - `phpstan/phpstan-symfony`
+
 ## [4.0.0] - 2025-06-24
 
 ### Feature